Spotbugs
-fixed bug

Author: @kennbr, @julianao, @sondrkol

Relates: #56

+++ passwordManager/core/src/test/java/core/UserSessionTest.java
@@ -3,6 +3,7 @@ package core;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.io.File;
+import java.io.IOException;
 import java.util.ArrayList;
 
 import org.junit.jupiter.api.Test;
@@ -11,18 +12,35 @@ public class UserSessionTest {
   UserSession userSession = UserSession.getInstance();
   File file = new File("../localpersistence/src/resources/localpersistance/TestUsers.json");
 
-  public UserSessionTest() {
-    this.initDatabase();
-  }
-
-  public void initDatabase() {
-    file.delete();
+  private void resetFile() {
+    if (file.exists()) {
+      file.delete();
+    }
+    File jsonFile = new File(file.toString());
+    try {
+      jsonFile.createNewFile();
+    } catch (IOException e) {
+      e.printStackTrace();
+    }
     userSession.overridePath(file.toString());
   }
 
-  protected void finalize() {
-    file.delete();
-  }
+
+
+  //public UserSessionTest() {
+  //  this.initDatabase();
+  //}
+//
+  //public void initDatabase() {
+  //  file.delete();
+  //  userSession.overridePath(file.toString());
+  //}
+//
+  //protected void finalize() {
+  //  file.delete();
+  //}
+
+
 
   private boolean hasProfile(ArrayList<Profile> profiles, Profile profile) {
     for (Profile p : profiles) {
@@ -38,7 +56,8 @@ public class UserSessionTest {
 
   @Test
   public void insertProfileTest() {
-    this.initDatabase();
+    //this.initDatabase();
+    resetFile();
     userSession.getDatabaseTalker().insertUser(new User("Admin", "Admin1!"));
 
     userSession.login("Admin", "Admin1!");
@@ -55,7 +74,8 @@ public class UserSessionTest {
 
   @Test
   public void loginTest() {
-    this.initDatabase();
+   // this.initDatabase();
+    resetFile();
     User user1 = new User("User1", "password1");
     User user2 = new User("User2", "password2");
 
@@ -72,7 +92,9 @@ public class UserSessionTest {
 
   @Test
   public void registerUserTest() {
-    this.initDatabase();
+    //this.initDatabase();
+    resetFile();
+
     assertEquals(false, userSession.login("user1", "password1"));
     userSession.registerUser("user1", "password1");
     assertEquals(true, userSession.login("user1", "password1"));
@@ -81,8 +103,9 @@ public class UserSessionTest {
 
   @Test
   public void userValidatorTest() {
-
-    this.initDatabase();
+   // this.initDatabase();
+    resetFile();
+    
     userSession.registerUser("user1", "Password1");
 
     String[][] tests = {
